requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-limiter==3.5.0
pymongo==4.6.1
passlib==1.7.4
certifi==2024.2.2
python-dotenv==1.0.1
